# Cleanup custom dictionary implementations