--- a/Jint/GlobalUsings.cs
+++ b/Jint/GlobalUsings.cs
@@ -1,0 +1,2 @@
+global using PropertyDictionary = Jint.Collections.HybridDictionary<Jint.Runtime.Descriptors.PropertyDescriptor>;
+global using SymbolDictionary = Jint.Collections.DictionarySlim<Jint.Native.JsSymbol, Jint.Runtime.Descriptors.PropertyDescriptor>;