… Jint/Collections/PropertyDictionary.cs
… Jint/Collections/SymbolDictionary.cs