Simplifying and adding idiomatic `GetValueRefOrNullRef` and `GetValueRefOrAddDefault`. This paves ground for further things.