@lahma 2025-01-24
> Simplifying -> `AddDangerous` ... But I trust you

Hah, it just for occasions when you know that thing is going to be added and no need to update/check existing value.